Revert "Rework thumbnail handling"
This reverts commit a1c6d0cc537f5d821034af2f0edc3a8cc04f0ee6.

Revert "Another optimization of glyph cache uploads"

This reverts commit fc786afbe0decf51d8a36f99e6f6821d0a0803be.

Revert "convert jint to uint32_t."

This reverts commit a8efed097230a019df82a0c9cafffcecd83a5d5b.

Revert "Fix regression in ListView click handling"

This reverts commit 6ee4b3ad6d56efc793bc03359e2678fece4d22df.

Revert "resolver list is not cleared on rebuild"

This reverts commit eff460a956e38ab0fb6f512731389cfba4ebbfa3.

Revert "StringBlock_nativeCreate: Possible Leak"

This reverts commit b1b5c5c0cf08d2056e78bba747c8b86d77ae7ef7.

Revert "Leak in ToneGenerator_native_setup"

This reverts commit 59c55d30c45dfb304ef14e6c78afd08b4823ae58.

Revert "Fix bugs regarding IllegalStateException during moving position in cursorwindow."

This reverts commit 3eee7775f083866d432bd3328a0e2a95cb4c0811.

Revert "Remove the fixed-length buffer in DdmHandleNativeHeap.getLeakInfo."

This reverts commit d8f135f2aaf5ae63c6e3f25c2877e51d42d62747.

Revert "Recognize odex files in memory accounting."

This reverts commit 31fd8af097a9892bcd781282bb53a9375a836555.

Revert "Account malloc mmaped pages as native heap."

This reverts commit df2121ca577e15690fb79abd4d4f7c7c83c2ad04.

Revert "Remove possible leak when exception occurred."

This reverts commit 9750dfa239fc9e50fded975f36bec59e926633b9.

Revert "Fix for breakText returning count of glyphs not chars."

This reverts commit e9b9ad2c5dab11512eda3c3b196fbcbda99ec328.

Revert "Track stack memory in MemoryInfo."

This reverts commit 975bf437386f58c6988918621241f72e1e500cfc.

Revert "Fix unexpected rotation change when re-enabling auto-rotate"

This reverts commit 04eebc0eba632ea0b33338a4180b401c0e8b7b30.

Revert "Function uri.getAuthority is called twice. Minor doc corrections."

This reverts commit b1d7f2c2d75a49271421b4bf1589f910b390dec7.

Revert "Patchset update."

This reverts commit de53dc367c808d304bf6337b359240a9bb7054d0.

Revert "libandroidfw: fix compilation warnings"

This reverts commit 94bd351810dc00e0eae3822672daaf82d15d73b7.

Revert "Disable usage of WAL in DatabaseHelper"

This reverts commit 7ab1dd616127894b6c49df95eac07857f8d78f17.

Revert "add reply.writeNoException() for quick response"

This reverts commit d64749ff2bb4ba9708bebf36530d4befc7e71deb.

Revert "fix hover events consume issue"

This reverts commit 51ef38bfef5a34786a5b5f676335470c0208542b.

Revert "Fix PointerCoords Parcel incorrectly storing values as Int32"

This reverts commit c90bb6d808eb1e9dfcf949d15f25a4da18d720cc.

Revert "Fixed the problem ime invisible status despite being the icon that appears in the statusbar."

This reverts commit 3c61d28c7c9ffca50342aa0fbbd4e028bd2fe860.

Revert "Use getRelativeDayString in getRelativeTimeSpanString."

This reverts commit 573c314cd65aac16d149c4d6240ee05381862ba7.

Revert "Option to set notification icon opacity 1/2"

This reverts commit 5e4c068b71142f20c45088a57fa9284d809f6bb8.

Revert ""[+>" more icon was never shown in PhoneStatusBar"

This reverts commit 3193f3374c70a600a673a07026363433b7333045.

Revert "Fix race condition in ActivityThread.completeRemoveProvider()"

This reverts commit 34f5e48698d7caa9547d13613f2c2806a8d64aa1.

Revert "Correct the target app token for input method window"

This reverts commit fb551d5af9a100de236d817b260ce0b7fd846a3e.

Revert "Handle nested folder case"

This reverts commit 2e3e0c3a493aec4bea60d4195a4290e1bd08001b.

Revert "Fix NullpointException problem in onMenuItemSelected"

This reverts commit 87c6077b34c9f968e2ee987f0c533e488463c50f.

Revert "Fix EventLog string class problem in onOptionMenuSelected"

This reverts commit be953eadf6e1a16665e843cc973fecc67be08a6f.

Revert "Fix ViewRootImpl to find missing focus using D-pad."

This reverts commit 2997b703fe8eebe5b5bd43c4cc247dbcda0d943d.

Revert "Make implementation of isEmpty consistent with implementation of getCount in HeaderListViewAdapter"

This reverts commit 33ad1a490d5b5752b24a48a5a35a6886f5a94b26.

Revert "Ensure running mTouchModeReset when AbsListView is detached"

This reverts commit 1e90714839127cd66b26e3244af5a5cf3d2c8dfb.

Revert "Make transactGetSockaddr locale safe"

This reverts commit 0a5c0ce350a5aa3db6106c27056895884fd84633.

Revert "show SELECT ALL icon with text in landscape mode"

This reverts commit b3ffd99bbf7b3240cd9395b50e172e4410cbc10f.

Revert "Two REMOVE_PROVIDER messages caused by race condition."

This reverts commit 76bbc72a701d4003590b23dad12e99b98be8b234.

Revert "Setting max hidden apps to 55 "

This reverts commit 3417fd5f8c83bbf1f696b3f248ab672c49b71362.

Revert "activity manager tweak-a-roonie"

This reverts commit 967b42fb4433cdb001e304c23d7d6ca369c860d3.

Revert "just a lil tweakaroonie to make recents draw with hardware acceleration"

This reverts commit 7f6e46f5763730187b5a6c4ea5851008f8da4234.

Revert "Fixed incorrect calculation of dirty invalidation area"

This reverts commit 13f0ff375674cf2b3d717b68619c23d0d7264e10.

Revert "FULL High compression and png optimization 33% lossless reduction. Bring It"

This reverts commit "53224f7d26bdb28cfbce165f0d9cf2b686438826"

Fix image crash

+++ libs/androidfw/Input.cpp
@@ -221,7 +221,7 @@ status_t PointerCoords::readFromParcel(Parcel* parcel) {
     }
 
     for (uint32_t i = 0; i < count; i++) {
-        values[i] = parcel->readFloat();
+        values[i] = parcel->readInt32();
     }
     return OK;
 }
@@ -231,7 +231,7 @@ status_t PointerCoords::writeToParcel(Parcel* parcel) const {
 
     uint32_t count = __builtin_popcountll(bits);
     for (uint32_t i = 0; i < count; i++) {
-        parcel->writeFloat(values[i]);
+        parcel->writeInt32(values[i]);
     }
     return OK;
 }
